+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.0.0-beta.1
+
+- Support for `fields` in forcible.rest.sObjectRows()
+- Support for CRUD in forcible.rest.sObjectQuickActions()
+- Support for forcible.rest.quickActions()
+- Add update record sample in README
+
 ## 1.0.0-beta.0
 
 - Support for CRUD operations